backend/src/ft_transcendence/api/serializers.py
from .models import User
from rest_framework import serializers

class UserSerializer(serializers.HyperlinkedModelSerializer):
	class Meta:
		model = User
		fields = ['url', 'username', 'displayName', 'avatar', 'date_joined', 'wins', 'looses']